Extract FieldOptionsList into its own component
